dagster-global-gf/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "dagster_global_gf"

# Main project configuration (dagster_global_gf)
[project]
name = "dagster_global_gf"
version = "0.9.0"
description = "dagster_global_gf for GF"
authors = [{ name = "Axell Padilla", email = "brian.padilla@farinter.com" }]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
# Development Status :: 1 - Planning
# Development Status :: 2 - Pre-Alpha
# Development Status :: 3 - Alpha
# Development Status :: 4 - Beta
# Development Status :: 5 - Production/Stable
# Development Status :: 6 - Mature
# Development Status :: 7 - Inactive
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Assets Orchestration",
    "Topic :: Database",
    "Topic :: Data Processing",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Framework :: Dagster",
]

dependencies = [
    "dagster_shared_gf",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock", "mypy", "ruff", "pre-commit"]

[tool.setuptools]
packages = ["dagster_global_gf"]

[tool.setuptools.exclude-package-data]
dagster_global_gf = ["dagster_global_gf_tests"]

[project.urls]
Repository = "https://github.com/farinter-analitica/main-dagster"
